requirements.txt
Flask==2.0.1
openpyxl==3.0.9
python-pptx==0.6.21
Werkzeug==2.0.1